added navbar and updated project directory with routes folder
