+++ frontend/src/routes/Home.js
@@ -1,10 +1,10 @@
-import { Link } from "react-router-dom";
+import CommonNavbar from "../components/navbar";
 
 const Home = () => {
   return (
     <div>
+      <CommonNavbar />
       <h1>Welcome to Home</h1>
-      <Link to="/form">Go to Form</Link>
     </div>
   );
 };
